Remove section to simplify explainer
@cwilso gave feedback that both the section on killing tasklets as well as the section on JavaScript integration are too into the weeds. This PR removes them.

Like @cwilso, I also felt that this section feels like a step 3 while we are working on step 1. Additionally, we might want to have a similar syntax for ServiceWorkers and other worklets in the future, so might be worth putting it into a separate proposal alltogether.

I think we can just state that tasklets have same behavior as Workers. No need to justify it. If anyone asks, we can have this discussion in the open.